+++ packages/blade/src/components/Checkbox/checkboxTokens.ts
@@ -1,4 +1,5 @@
 import type { Theme } from '~components/BladeProvider';
+import type { SelectorInputHoverTokens } from '~components/Form/Selector/types';
 import type { DotNotationColorStringToken } from '~src/_helpers/types';
 import { size } from '~tokens/global';
 
@@ -82,4 +83,17 @@ const checkboxIconColors: CheckboxIconColors = {
   },
 } as const;
 
-export { checkboxSizes, checkboxIconColors };
+const checkboxHoverTokens: SelectorInputHoverTokens = {
+  default: {
+    background: {
+      checked: 'colors.brand.primary.600',
+      unchecked: 'colors.brand.gray.a50.lowContrast',
+    },
+    border: {
+      checked: 'colors.brand.primary.600',
+      unchecked: 'colors.brand.gray.500.lowContrast',
+    },
+  },
+};
+
+export { checkboxSizes, checkboxIconColors, checkboxHoverTokens };